requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
appdirs==1.4.4            # via virtualenv
asgiref==3.2.10           # via django
attrs==20.2.0             # via pytest
cfgv==3.2.0               # via pre-commit
distlib==0.3.1            # via virtualenv
django==3.1.1             # via -r requirements.in, django-localflavor, django-phonenumber-field, djangorestframework, djangorestframework-simplejwt
django-extensions==3.0.9  # via -r requirements.in
django-localflavor==3.0.1  # via -r requirements.in
django-phonenumber-field==5.0.0  # via -r requirements.in
django-phonenumbers==1.0.1  # via -r requirements.in
django-rest-framework==0.1.0  # via -r requirements.in
djangorestframework==3.12.1  # via django-rest-framework, djangorestframework-simplejwt
djangorestframework-simplejwt==4.4.0  # via -r requirements.in
faker==4.5.0              # via -r requirements.in
filelock==3.0.12          # via virtualenv
identify==1.5.5           # via pre-commit
iniconfig==1.0.1          # via pytest
nodeenv==1.5.0            # via pre-commit
packaging==20.4           # via pytest
phonenumbers==8.12.10     # via django-phonenumbers
pluggy==0.13.1            # via pytest
pre-commit==2.7.1         # via -r requirements.in
psycopg2-binary==2.8.6    # via -r requirements.in
py==1.9.0                 # via pytest
pyjwt==1.7.1              # via djangorestframework-simplejwt
pyparsing==2.4.7          # via packaging
pytest==6.1.0             # via pytest-django
pytest-django==3.10.0     # via -r requirements.in
python-dateutil==2.8.1    # via faker
python-stdnum==1.14       # via django-localflavor
pytz==2020.1              # via django
pyyaml==5.3.1             # via pre-commit
six==1.15.0               # via packaging, python-dateutil, virtualenv
sqlparse==0.3.1           # via django
text-unidecode==1.3       # via faker
toml==0.10.1              # via pre-commit, pytest
virtualenv==20.0.32       # via pre-commit
